# swagger.json
{
	"swagger": "2.0",
	"info": {
		"version": "1.0",
		"title": "final-project",
		"contact": {}
	},
	"host": "localhost:8888",
	"basePath": "/",
	"securityDefinitions": {
		"Bearer": {
			"type": "apiKey",
			"name": "Authorization",
			"in": "header",
			"description": "Enter your bearer token in the format **Bearer &lt;token>**"
		}
	},
	"schemes": [
		"http"
	],
	"consumes": [
		"application/json"
	],
	"produces": [
		"application/json"
	],
	"paths": {
...
		"/auth/me": {
			"get": {
				"summary": "Current User",
				"tags": [
					"Auth"
				],
				"security": [{
					"Bearer": []
				}],
				"operationId": "CurrentUser",
				"deprecated": false,
				"produces": [
					"application/json"
				],
				"parameters": [],
				"responses": {
					"200": {
						"description": "",
						"headers": {}
					}
				}
			}
		},
...
		"/api/transaction/{id}": {
			"get": {
				"summary": "getTransactionByUserId",
				"tags": [
					"getTransaction"
				],
				"security": [{
					"Bearer": []
				}],
				"operationId": "getTransactionByUserId",
				"deprecated": false,
				"produces": [
					"application/json"
				],
				"parameters": [{
					"name": "id",
					"in": "path",
					"required": true,
					"type": "string",
					"description": ""
				}],
				"responses": {
					"200": {
						"description": "",
						"headers": {}
					}
				}
			}
		},
		"/api/transactionOwner/{id}": {
			"get": {
				"summary": "getTransactionByOwnerId",
				"tags": [
					"getTransaction"
				],
				"security": [{
					"Bearer": []
				}],
				"operationId": "getTransactionByOwnerId",
				"deprecated": false,
				"produces": [
					"application/json"
				],
				"parameters": [{
					"name": "id",
					"in": "path",
					"required": true,
					"type": "string",
					"description": ""
				}],
				"responses": {
					"200": {
						"description": "",
						"headers": {}
					}
				}
			}
		},
		"/api/transaction": {
			"post": {
				"summary": "create transaction",
				"tags": [
					"CRUD transaction"
				],
				"security": [{
					"Bearer": []
				}],
				"operationId": "createtransaction",
				"deprecated": false,
				"produces": [
					"application/json"
				],
				"parameters": [{
					"name": "Body",
					"in": "body",
					"required": true,
					"description": "",
					"schema": {
						"$ref": "#/definitions/createtransactionrequest"
					}
				}],
				"responses": {
					"200": {
						"description": "",
						"headers": {}
					}
				}
			}
		}
	},
	"definitions": {
...
		"createtransactionrequest": {
			"title": "createtransactionrequest",
			"example": {
				"product_id": 1,
				"owner_id": 1,
				"requestedPrice": 200000,
				"isOpened": true
			},
			"type": "object",
			"properties": {
				"product_id": {
					"type": "integer",
					"format": "int32"
				},
				"owner_id": {
					"type": "integer",
					"format": "int32"
				},
				"requestedPrice": {
					"type": "integer",
					"format": "int32"
				},
				"isOpened": {
					"type": "boolean"
				}
			},
			"required": [
				"product_id",
				"owner_id",
				"requestedPrice",
				"isOpened"
			]
		},
...
	},
	"security": [],
...
}